Move SPI files into SPI module
