+++ .env
@@ -1,2 +1,2 @@
-TAG=0.28.1-SNAPSHOT
+TAG=0.32.0-SNAPSHOT
 REGISTRY_PREFIX=
