--- hedera-node/hedera-app-api/src/main/java/module-info.java
@@ -1,3 +1,0 @@
-module com.hedera.services.api {
-    exports com.hedera.services.api;
-}
