+++ hedera-node/hedera-app-spi/src/main/java/com/hedera/node/app/spi/package-info.java
@@ -5,4 +5,4 @@
  *
  * <p><b>There should be no implementation code inside this package.</b>
  */
-package com.hedera.services.api;
+package com.hedera.node.app.spi;
