+++ hedera-node/hedera-app/src/main/java/com/hedera/services/api/implementation/PlaceholderImplementation.java
@@ -15,11 +15,9 @@
  */
 package com.hedera.services.api.implementation;
 
-import com.hedera.services.api.PlaceholderApi;
-
 /**
  * Placeholder <b>implementation</b> for to-be-determined public APIs. We'll remove this class when
- * we remove the {@link PlaceholderApi} interface.
+ * we add real implementation code.
  */
 @SuppressWarnings("unused")
-public class PlaceholderImplementation implements PlaceholderApi {}
+public class PlaceholderImplementation {}
