+++ hedera-node/hedera-app/src/main/java/module-info.java
@@ -1,3 +1,1 @@
-module com.hedera.services.api.implementation {
-    requires com.hedera.services.api;
-}
+module com.hedera.services.api.implementation {}
