+++ hedera-node/hedera-mono-service/build.gradle.kts
@@ -24,6 +24,7 @@ dependencies {
     annotationProcessor(libs.dagger.compiler)
 
     api(project(":hedera-node:hedera-evm-api"))
+    api(project(":hedera-node:hedera-app-spi"))
     implementation(project(":hapi-fees"))
     implementation(project(":hapi-utils"))
     implementation(libs.bundles.besu) {
