+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/node/app/service/token/impl/CryptoPreTransactionHandlerImplTest.java
@@ -22,7 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.BDDMockito.given;
 
-import com.hedera.node.app.spi.meta.SigTransactionMetadata;
+import com.hedera.node.app.SigTransactionMetadata;
 import com.hedera.node.app.spi.state.States;
 import com.hedera.node.app.state.impl.InMemoryStateImpl;
 import com.hedera.node.app.state.impl.RebuiltStateImpl;
