+++ settings.gradle.kts
@@ -24,7 +24,7 @@ include(":hapi-utils")
 include(":hapi-fees")
 include(":hedera-node")
 include(":hedera-node:hedera-app")
-include(":hedera-node:hedera-app-api")
+include(":hedera-node:hedera-app-spi")
 include(":hedera-node:hedera-evm")
 include(":hedera-node:hedera-evm-api")
 include(":hedera-node:hedera-mono-service")
